Reorganize functions into files
